Refactorizacion todo menos Put
